backend/requirements.txt
# FastAPI
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Validation
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Telegram
python-telegram-bot==20.8

# Utilities
python-dateutil==2.8.2
aiofiles==23.2.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
httpx==0.26.0
pytest-cov==4.1.0

